requirements.txt
Flask==2.3.2
Flask_Cors==3.0.10
geopy==2.3.0
jmespath==1.0.1
ppygis3==0.4
psycopg2==2.9.6
python-dotenv==1.0.0
